# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SE Project API (Team 8, May 2025)
  description: "Backend API for Team 8's Software Engineering Project (May 2025 Term)\n\
    \nThis backend powers a web-based mobile application designed for senior citizens\
    \ and their family caregivers. The application enhances daily living and care\
    \ coordination through these core modules identified based on user stories:\n\n\
    - Easy-to-use UI: Senior-friendly design with large text, bold colors, and simple\
    \ navigation\n- Medicine Tracker: Prescription management and medication identification\n\
    - Appointments Tracker: Book, view, and get reminders for doctor appointments\n\
    - SOS: Emergency alert system for designated contacts\n- Documents Vault: Secure\
    \ digital repository for medical reports and history\n\n---\n\n## User Story Mapping\n\
    \n| # | User Story (Summary) | API Endpoints |\n|---|----------------------|--------------|\n\
    | 1 | UI accessibility | (Frontend/UI, not API) |\n| 2 | Medicine management |\
    \ `/medicines/`, `/medicines/{medicine_id}`, `/medicines/transcribe` |\n| 3 |\
    \ Appointment management | `/appointments/`, `/appointments/user/{user_id}` |\n\
    | 4 | Emergency contact notification | `/users/{user_id}/sos/trigger`, `/emergency-contacts/`\
    \ |\n| 5 | Document storage | `/documents/`, `/documents/upload`, `/documents/user/{user_id}`\
    \ |\n| 6 | Medication adherence tracking | `/medicines/`, `/medicines/user/{user_id}`\
...
paths:
...
  /api/v1/doctors/:
    post:
      tags:
      - Doctors
      summary: Create Doctor
      description: 'Create a new doctor record. Requires admin authentication. Clears
        doctors cache.


        Supports US9 and US11 by enabling management of doctor profiles and facilitating
        report sharing.'
      operationId: create_doctor_api_v1_doctors__post
      parameters:
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '201':
          description: Doctor created successfully.
        '400':
          description: Invalid input.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Doctors
      summary: Get All Doctors
      description: 'Get all doctors. Results are cached for 5 minutes.


        Supports US9 and US11 by providing access to doctor lists for users and families.'
      operationId: get_all_doctors_api_v1_doctors__get
      responses:
        '200':
          description: List of all doctors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorResponse'
                title: Response Get All Doctors Api V1 Doctors  Get
  /api/v1/doctors/{doctor_id}:
    get:
      tags:
      - Doctors
      summary: Get Doctor By Id
      description: 'Get a doctor by ID. Results are cached for 5 minutes.


        Supports US9 and US11 by letting users and families review doctor details
        for appointments and consultations.'
      operationId: get_doctor_by_id_api_v1_doctors__doctor_id__get
      parameters:
      - name: doctor_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the doctor to retrieve
          title: Doctor Id
        description: ID of the doctor to retrieve
      responses:
        '200':
          description: Doctor found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '404':
          description: Doctor not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Doctors
      summary: Update Doctor
      description: 'Update an existing doctor record. Requires admin authentication.
        Clears doctors cache.


        Supports US9 and US11 by keeping doctor information up to date for users and
        families.'
      operationId: update_doctor_api_v1_doctors__doctor_id__put
      parameters:
      - name: doctor_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the doctor to update
          title: Doctor Id
        description: ID of the doctor to update
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorUpdate'
      responses:
        '200':
          description: Doctor updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '404':
          description: Doctor not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Doctors
      summary: Delete Doctor
      description: 'Delete a doctor by ID. Requires admin authentication. Clears doctors
        cache.


        Supports US9 and US11 by allowing removal of outdated or incorrect doctor
        records.'
      operationId: delete_doctor_api_v1_doctors__doctor_id__delete
      parameters:
      - name: doctor_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the doctor to delete
          title: Doctor Id
        description: ID of the doctor to delete
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Doctor deleted successfully.
          content:
            application/json:
              schema: {}
        '404':
          description: Doctor not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/appointments/doctor/{doctor_id}:
    get:
      tags:
      - Appointments
      summary: Get Appointments By Doctor
      description: 'List all appointments for a doctor. Results are cached for 5 minutes.


        Supports US7 by enabling doctors and families to coordinate appointments.'
      operationId: get_appointments_by_doctor_api_v1_appointments_doctor__doctor_id__get
      parameters:
      - name: doctor_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the doctor to get appointments for
          title: Doctor Id
        description: ID of the doctor to get appointments for
      responses:
        '200':
          description: List of appointments for the doctor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'
                title: Response Get Appointments By Doctor Api V1 Appointments Doctor  Doctor
                  Id  Get
        '404':
          description: Doctor not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AppointmentResponse:
      properties:
        name:
          type: string
          title: Name
          description: Appointment name
          example: Consultation
        date:
          type: string
          format: date
          title: Date
          description: Appointment date (YYYY-MM-DD)
          example: '2025-07-03'
        time:
          type: string
          format: time
          title: Time
          description: Appointment time (HH:MM:SS)
          example: '14:30:00'
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Additional notes (optional)
          example: Eat healthy
        id:
          type: integer
          title: Id
          description: Unique appointment ID
          example: 1
        user_id:
          type: integer
          title: User Id
          description: User ID for the appointment
          example: 1
        doctor_id:
          type: integer
          title: Doctor Id
          description: Doctor ID for the appointment
          example: 1
        medicines:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Medicines
          description: List of medicine IDs for this appointment (optional)
      type: object
      required:
      - name
      - date
      - time
      - id
      - user_id
      - doctor_id
      title: AppointmentResponse
      description: Schema for returning appointment data to the client.
      example:
        date: '2025-07-03'
        doctor_id: 2
        id: 1
        medicines:
        - 10
        - 11
        name: Consultation
        notes: Eat healthy
        time: '14:30:00'
        user_id: 1
...
    DoctorCreate:
      properties:
        name:
          type: string
          title: Name
          description: Doctor's name
          example: Dr. A. Kumar
        location:
          type: string
          title: Location
          description: Doctor's location
          example: Delhi
      type: object
      required:
      - name
      - location
      title: DoctorCreate
      description: Schema for creating a new doctor record.
    DoctorResponse:
      properties:
        name:
          type: string
          title: Name
          description: Doctor's name
          example: Dr. A. Kumar
        location:
          type: string
          title: Location
          description: Doctor's location
          example: Delhi
        id:
          type: integer
          title: Id
          description: Unique doctor ID
          example: 1
      type: object
      required:
      - name
      - location
      - id
      title: DoctorResponse
      description: Schema for returning doctor data to the client.
      example:
        id: 1
        location: Delhi
        name: Dr. A. Kumar
    DoctorUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Doctor's name (optional)
          example: Dr. A. Kumar
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
          description: Doctor's location (optional)
          example: Delhi
      type: object
      title: DoctorUpdate
      description: Schema for updating a doctor record (partial update).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError